pyproject.toml
[project]
name = "snr-intelligence"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.136.0",
    "passlib>=1.7.4",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.49",
    "sqlmodel>=0.0.38",
    "uvicorn[standard]>=0.44.0",
]
